fix travel preference save; fix DB syntax error; refactor gateway function and logging;

+++ src/ACMESkyService/entities/flight_subscription.go
@@ -8,7 +8,7 @@ type CustomerFlightSubscription struct {
 	DateStartISO8601 string `json:"travel_date_start"`
 	// end travel date range in ISO 8601 format (with timezone UTC)
 	DateEndISO8601 string  `json:"travel_date_end"`
-	Budget         float32 `json:"travel_max_price"`
+	Budget         float64 `json:"travel_max_price"`
 }
 
 func CustomerFlightSubscriptionFromMap(m map[string]interface{}) CustomerFlightSubscription {
@@ -18,7 +18,7 @@ func CustomerFlightSubscriptionFromMap(m map[string]interface{}) CustomerFlightS
 		AirportDestinationID: m["airport_id_destination"].(string),
 		DateStartISO8601:     m["travel_date_start"].(string),
 		DateEndISO8601:       m["travel_date_end"].(string),
-		Budget:               m["travel_max_price"].(float32),
+		Budget:               m["travel_max_price"].(float64),
 	}
 	return v
 }
